Remove all-of-dublin stats
This has become deprecated as newer data has come to light, as it was
too difficult to add to and run.

It is also too complex for such a simple job as summing up a few values
into a pie chart.  Best to keep individual site operations in Python
with a workflow for each dataset and perform cross-dataset operations
in a simpler tool