Add Binder, GitPod & Local setup instructions
Simplify GitPod by installing all project dependencies on
startup via micromamba.  Am having trouble adding micromamba
init script to the ~/.bashrc as it seems to be overwritten when
re-launching a workspace.  This isn't an issue if a new workspace
is launched every time, however, in this case ~350 MB of libraries
are re-installed every time...