Fill unknown buildings with archetypes & save
Instead of keeping buildings in memory, loading archetypes
and filling in a for loop, filter out unknown buildings and
fill these with archetypes.  This implementation is messy
but does overcome binder memory issues.  Ideally the logic
would treat dataframes as immutable and map archetypes to
unknown buildings creating a new dataframe.

Closes #125